Redux start

+++ server/app.js
@@ -63,14 +63,15 @@ app.post('/login', (req, res) => {
     } else {
       if (row.length) {
         if (bcrypt.compareSync(password, row[0].password)) {
-          session.userId = row[0].id
+          let sessionId = row[0].id;
+          // session.userId = row[0].id
           res.cookie('session-cookie', 'loggedIn', { maxAge: 180000 });
           res.status(200).json({ userId: row[0].id, twofactor: row[0].twofactor, password: row[0].password});
         } else {
           res.status(401).json({ error : "Incorrect password"});
         }
       } else {
-        res.status(401).json({ error : "Invalid username"});
+        // res.status(401).json({ error : "Invalid username"});
       }
     }
   });
@@ -97,7 +98,7 @@ app.get('/usernames', (req, res) => {
 })
 
 app.get('/profile', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
@@ -184,7 +185,7 @@ app.post('/signup', (req, res) => {
 
 app.post('/pay', (req, res) => {
   // TODO: check if user is still logged in (i.e. check cookie) here. If not, send back appropriate error response.
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
@@ -219,14 +220,12 @@ app.post('/pay', (req, res) => {
 
 
 app.get('/publicprofile', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
-
   // Refresh the cookie timer
   res.cookie('session-cookie', 'loggedIn', { maxAge: 180000 });
-
 
   let username = req.query.username;
   username = username && _.escape(username.replace(/"/g,"'"));
@@ -260,7 +259,7 @@ app.get('/publicprofile', (req, res) => {
 const FEED_DEFAULT_LENGTH = 5;
 
 app.get('/feed/global', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
@@ -285,7 +284,7 @@ app.get('/feed/global', (req, res) => {
 });
 
 app.get('/feed/all', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
@@ -306,7 +305,7 @@ app.get('/feed/all', (req, res) => {
 });
 
 app.get('/feed/user/:userId', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
@@ -336,11 +335,10 @@ app.get('/feed/user/:userId', (req, res) => {
 });
 
 app.get('/feed/profile', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
-
   // Refresh the cookie timer
   res.cookie('session-cookie', 'loggedIn', { maxAge: 180000 });
 
@@ -365,11 +363,10 @@ app.get('/feed/profile', (req, res) => {
 });
 
 app.get('/feed/relational', (req, res) => {
-  if (!req.headers.cookie.includes('session-cookie=loggedIn')) {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
     res.status(403).json({ error: 'User logged out for security reasons'});
     return;
   }
-
   // Refresh the cookie timer
   res.cookie('session-cookie', 'loggedIn', { maxAge: 180000 });
 
@@ -396,6 +393,13 @@ app.get('*', (req, res) => {
 });
 
 app.patch('/request', (req, res) => {
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
+    res.status(403).json({ error: 'User logged out for security reasons'});
+    return;
+  }
+  // Refresh the cookie timer
+  res.cookie('session-cookie', 'loggedIn', { maxAge: 180000 });
+
   db.updateRequest(req.body.transactionId)
     .then((results) => {
       res.status(201).json();
@@ -406,7 +410,13 @@ app.patch('/request', (req, res) => {
 });
 
 app.post('/request', (req, res) => {
-  console.log(req.body);
+  if (!req.headers.cookie || !req.headers.cookie.includes('session-cookie=loggedIn')) {
+    res.status(403).json({ error: 'User logged out for security reasons'});
+    return;
+  }
+  // Refresh the cookie timer
+  res.cookie('session-cookie', 'loggedIn', { maxAge: 180000 });
+
   db.deleteRequest(req.body.transactionId)
     .then((results) => {
       res.status(200).json();
@@ -450,4 +460,5 @@ router.get("/invite", function(req, res) {
 // https://app.sendgrid.com/guide/integrate/langs/nodejs/verify
 
 //==================================//
+
 module.exports = app;
